RVN-7866 - завершил рефакторинг парсинга, надо тестировать
