+++ package.json
@@ -20,6 +20,7 @@
     "csv-parse": "2.4.1",
     "ip": "1.1.5",
     "lodash": "3.10.1",
+    "moment": "^2.23.0",
     "xml-stream": "0.4.5"
   },
   "devDependencies": {
